content-service/etl/requirements.txt
backoff==2.1.2
certifi==2022.6.15
elastic-transport==8.1.2
elasticsearch==7.0.0
flake8==4.0.0
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
pyflakes==2.4.0
python-dotenv==0.20.0
typing_extensions==4.3.0
urllib3==1.26.9
